Move generate_gaussian_noise to new utils.py file
